pyproject.toml
[build-system]
requires = [
    "setuptools>=45",
    "setuptools-scm[toml]>=6.2",
    "wheel>=0.38.0",
    "importlib_resources>=1.3",
    "grpcio-tools>=1.48.2,<=1.54.0",
    "googleapis-common-protos>=1.3.1,<=1.59.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "mtap"
description = "A framework for distributed text analysis using gRPC and microservices-based architecture."
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE.txt"}
keywords = ["nlp", "grpc"]
authors = [
    {name = "University of Minnesota NLP/IE Group", email = "nlp-ie@umn.edu"},
    {name = "Ben Knoll", email = "benknoll@umn.edu"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Java",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Java Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "grpcio>=1.48.2,<=1.54.0",
    "grpcio-health-checking>=1.48.2,<=1.54.0",
    "grpcio-status>=1.48.2,<=1.54.0",
    "pyyaml>=6.0,<=6.0",
    "tqdm>=4.0.0,<=4.65.0",
    "googleapis-common-protos>=1.3.1,<=1.59.0",
    "aiofiles>=23.1.0,<=23.1.0"
]
dynamic = ["version"]

[tool.setuptools.packages.find]
where = ["python"]
include = ["mtap*"]

[tool.setuptools_scm]
write_to = "python/mtap/version.py"

[project.optional-dependencies]
test = [
    "pytest==7.3.1",
    "pytest-mock==3.10.0",
    "grpcio-testing>=1.48.2,<=1.54.0",
    "requests==2.28.2",
    "importlib-resources==5.12.0",
]
docs = [
    "sphinx==6.2.0",
]
consul = ["python-consul"]

[project.urls]
homepage = "https://nlpie.github.io/mtap"
documentation = "https://nlpie.github.io/mtap-python-api/index.html"
repository = "https://github.com/nlpie/mtap.git"
